♻️ refactor: improve Storybook configuration
