--- frontend/apps/app/.storybook/preview.ts
@@ -1,17 +1,0 @@
-import type { Preview } from '@storybook/react'
-import '../app/globals.css'
-import './langfuseWeb.mock'
-
-const preview: Preview = {
-  parameters: {
-    actions: { argTypesRegex: '^on[A-Z].*' },
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/,
-      },
-    },
-  },
-}
-
-export default preview
