--- frontend/apps/app/.storybook/vitest.setup.ts
@@ -1,9 +1,0 @@
-import { setProjectAnnotations } from '@storybook/react'
-import { beforeAll } from 'vitest'
-import * as projectAnnotations from './preview'
-
-// This is an important step to apply the right configuration when testing your stories.
-// More info at: https://storybook.js.org/docs/api/portable-stories/portable-stories-vitest#setprojectannotations
-const project = setProjectAnnotations([projectAnnotations])
-
-beforeAll(project.beforeAll)
