+++ frontend/apps/app/package.json
@@ -51,11 +51,6 @@
     "@liam-hq/configs": "workspace:*",
     "@octokit/openapi-types": "25.0.0",
     "@prisma/nextjs-monorepo-workaround-plugin": "6.6.0",
-    "@storybook/addon-essentials": "8.6.12",
-    "@storybook/addon-interactions": "8.6.12",
-    "@storybook/addon-links": "8.6.12",
-    "@storybook/blocks": "8.6.12",
-    "@storybook/nextjs": "8.6.12",
     "@storybook/react": "8.6.12",
     "@types/diff": "7.0.2",
     "@types/gtag.js": "0.0.20",
@@ -71,7 +66,6 @@
   },
   "scripts": {
     "build": "pnpm gen && next build",
-    "build-storybook": "storybook build",
     "dev": "concurrently \"pnpm:dev:*\"",
     "dev:css": "tcm -p '!(node_modules|.next|.trigger|.turbo)/**/*.module.css' --watch",
     "dev:next": "next dev --port 3001",
@@ -86,7 +80,6 @@
     "lint:tsc": "tsc --noEmit",
     "postinstall": "cp ../../packages/db-structure/node_modules/@ruby/prism/src/prism.wasm prism.wasm",
     "start": "next start",
-    "storybook": "storybook dev -p 6006",
     "test": "vitest --watch=false",
     "test:vitest": "pnpm vitest"
   }
