+++ frontend/packages/ui/src/components/Button/Button.stories.tsx
@@ -1,5 +1,5 @@
-import { Button } from '@liam-hq/ui'
 import type { Meta, StoryObj } from '@storybook/react'
+import { Button } from './Button'
 
 const meta: Meta<typeof Button> = {
   title: 'UI/Button',
